chore: Update fail2ban configuration for Nextcloud (#15)
This commit updates the fail2ban configuration for Nextcloud. It removes the outdated 100-fail2ban-filter.conf and 100-fail2ban.conf.j2 files and replaces them with the new fail2ban-filter.conf and fail2ban-jail.conf.j2 files. The new configuration files have been renamed to improve clarity and consistency. This change ensures that the fail2ban configuration is up to date and aligned with the latest Nextcloud installation process.